[ALPS03091777] DISP: mt8167 hdmi plug in/out
[Detail]
when play secure video and plug out hdmi,
rdma1 will still be secure mode.

[Solution]
1. use trigger lock in hdmi_suspend.
2. change rdma1 to normal in hdmi_suspend.
3. change ext config and trigger as merge such as primary.

CR-Id: ALPS03091777
Feature: Secure Video Path (SVP)
MTK-Commit-Id: 4ce62e6c2196ed453c3f6c216b79405cf4562c54

+++ drivers/misc/mediatek/video/mt8167/dispsys/ddp_dsi.c
@@ -3438,7 +3438,7 @@ const LCM_UTIL_FUNCS PM_lcm_utils_dsi0 = {
 
 void *get_dsi_params_handle(uint32_t dsi_idx)
 {
-	if (dsi_idx != PM_DSI1)
+	if (dsi_idx != (uint32_t)(PM_DSI1))
 		return (void *)(&_dsi_context[0].dsi_params);
 	else
 		return (void *)(&_dsi_context[0].dsi_params);
